Dockerfile changes, installed more stuff. Changes in the AST format and the code to make the go testing happier
- Exported Python AST now working with the SDK (most fixes were in the pydetector depedency)
- Installed bash in the Dockerfiles
- Really read line by line + docker/makefile changes
- Python's readlines wait for the full file to complete. Changed it to
just iterate over the file object that by default does it by line
(without waiting for it to complete).
- Removed the checks for "language" and "language_version" in the
driver since it seems they're not in the protocol anymore and thus not
being send by the Go driver.
- Dockerfile / Makefile:
Added a sh that is installed to $(BUILD_PATH)/native just like the Java
driver do (this will be installed on /opt/driver/bin/native on the
container).
- Removed /opt/python_driver after it is installed by pip. Added
- Vim for easier debugging (will be removed before the final version).